muitas coisas

+++ programa/func.php
@@ -4,11 +4,27 @@
     require('models/Model.php');
     require('models/Documento.php');
     require('models/Compartilhamento.php');
-
-    $id = $_GET['view'] ?? false;    
-        $comp = new Compartilhamento();
-        $comp->delete( $id);
+    $pag = $_GET['pag'] ?? false;
+    $id = $_GET['view'] ?? false;   
+     
         $doc = new Documento();
-        $doc->delete($id);
-            header('location: documentos_lista.php');
-        
+        $comp = new Compartilhamento();
+        $documento = $doc->getById($id);
+        $compartilhamento = $comp->getByIdComp($id);
+
+
+        if($compartilhamento['nivel_per'] >= 3 or $_SESSION['user'] == $documento['usuarios_idusuarios']){
+               $comp->delete( $id);
+               $doc->delete($id);
+               header("location: documentos_$pag.php");
+    }else{
+        header("location: documentos_$pag.php");
+    }
+
+
+
+
+
+   
+
+
